[asm-cherry-pick] Fill exception message for max String literal length
Gatling compiles Scala classes on the fly with Zinc.

Be it a program bug or a user error, some Strings sometimes end up
being longer than the maximum String length. It would be nice if we
could get an Exception with the proper message.

I guess there's the same issue with maximum file size and maximum
method size, will check.