[asm-cherry-pick] expose Label.flags
And widen it to an int, as it was in previous versions of
ASM.

In order to efficiently implement removal of unreachable code:

https://github.com/scala/scala/blob/d8f5f34ca7461f38bc2466208bf2a87a4644aace/src/compiler/scala/tools/nsc/backend/jvm/opt/LocalOpt.scala#L495-L546

We stashed the reachablity boolean in a flag on Label, rather
than maintainig a `Map[Label, Boolean]`

We made that change in https://github.com/scala/scala/commit/6e3c5fcadaf03bfaff734067ae1475dc790b3a05

ASM 6.2 changed this field to a short. This commit restores it
to an int to be conservative (we know that our extra flags in
the high bits can't conflict with ASMs use of the field).